+++ ProjectTest/OneLang/Java/src/main/java/OneLang/StdLib/PackageManager/ImplPkgImplIntf.java
@@ -1,9 +1,9 @@
 public class ImplPkgImplIntf {
     public String name;
-    public Integer minver;
-    public Integer maxver;
+    public Double minver;
+    public Double maxver;
     
-    public ImplPkgImplIntf(String name, Integer minver, Integer maxver)
+    public ImplPkgImplIntf(String name, Double minver, Double maxver)
     {
         this.name = name;
         this.minver = minver;
